Bug 1788606 - part 5: Stop calling gradle in the decision task
