+++ taskcluster/android_taskgraph/build_config.py
@@ -8,7 +8,7 @@ import yaml
 
 from taskgraph.util.memoize import memoize
 
-from android_taskgraph import PROJECT_DIR, ANDROID_COMPONENTS_DIR
+from android_taskgraph import PROJECT_DIR, ANDROID_COMPONENTS_DIR, FOCUS_DIR
 
 
 EXTENSIONS = {
@@ -19,7 +19,7 @@ CHECKSUMS_EXTENSIONS = ('.sha1', '.md5')
 
 
 def get_components():
-    build_config = _read_build_config()
+    build_config = _read_build_config(ANDROID_COMPONENTS_DIR)
     return [{
         'name': name,
         'path': project['path'],
@@ -33,11 +33,11 @@ def get_version():
 
 
 def get_path(component):
-    return _read_build_config()["projects"][component]["path"]
+    return _read_build_config(ANDROID_COMPONENTS_DIR)["projects"][component]["path"]
 
 
 def get_extensions(component):
-    artifact_type = _read_build_config()["projects"][component].get("artifact-type", "aar")
+    artifact_type = _read_build_config(ANDROID_COMPONENTS_DIR)["projects"][component].get("artifact-type", "aar")
     if artifact_type not in EXTENSIONS:
         raise ValueError(
             "For '{}', 'artifact-type' must be one of {}".format(
@@ -53,6 +53,39 @@ def get_extensions(component):
 
 
 @memoize
-def _read_build_config():
-    with open(os.path.join(ANDROID_COMPONENTS_DIR, '.buildconfig.yml'), 'rb') as f:
+def _read_build_config(root_dir):
+    with open(os.path.join(root_dir, '.buildconfig.yml'), 'rb') as f:
         return yaml.safe_load(f)
+
+
+@memoize
+def get_upstream_deps_for_all_gradle_projects():
+    all_deps = {}
+    for root_dir in (ANDROID_COMPONENTS_DIR, FOCUS_DIR):
+        build_config = _read_build_config(root_dir)
+        all_deps.update({
+            project: project_config["upstream_dependencies"]
+            for project, project_config in build_config["projects"].items()
+        })
+
+    return all_deps
+
+
+def get_variant(build_type, build_name):
+    # TODO: Support Fenix
+    all_variants = _read_build_config(FOCUS_DIR)["variants"]
+    matching_variants = [
+        variant for variant in all_variants
+        if variant["build_type"] == build_type and variant["name"] == build_name
+    ]
+    number_of_matching_variants = len(matching_variants)
+    if number_of_matching_variants == 0:
+        raise ValueError('No variant found for build type "{}"'.format(
+            build_type
+        ))
+    elif number_of_matching_variants > 1:
+        raise ValueError('Too many variants found for build type "{}"": {}'.format(
+            build_type, matching_variants
+        ))
+
+    return matching_variants.pop()
