+++ taskcluster/android_taskgraph/transforms/gradle_optimization.py
@@ -1,7 +1,6 @@
 from taskgraph.transforms.base import TransformSequence
 
-from ..build_config import get_path
-from ..gradle import get_upstream_deps_for_all_gradle_projects
+from ..build_config import get_path, get_upstream_deps_for_all_gradle_projects
 
 
 transforms = TransformSequence()
